Edited the codes

+++ Practical/src/main/java/com/example/employeeService/dao/EmployeeDao.java
@@ -5,7 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.employeeService.model.Employee;
 
 
-public interface EmployeeDao extends JpaRepository<Employee, Long>{
+
+public interface EmployeeDao extends JpaRepository<Employee, Integer>{
+
+	//void save(Telephone telephone);
+
